+++ src/pages/ManagePage/StepSlope/components/ColumnFilterModal.tsx
@@ -1,12 +1,5 @@
 import styled from 'styled-components';
-import { Table } from '@tanstack/react-table';
-import { Slope } from '../../../../apis/slopeMap';
-
-interface FilterModalProps {
-  isOpen: boolean;
-  onClose: () => void;
-  table: Table<Slope>;
-}
+import { FilterModalProps } from '../../interface';
 
 const FilterModal = ({ isOpen, onClose, table }: FilterModalProps) => {
   return (
